Separate GoalTask class from simple navigation
